chore: fix `plus.*NonNumeric` errors from PHPStan (#9260)

+++ src/Fixer/ClassNotation/NoBlankLinesAfterClassOpeningFixer.php
@@ -91,7 +91,7 @@ final class NoBlankLinesAfterClassOpeningFixer extends AbstractFixer implements 
         // if there is more than one new line in the whitespace, then we need to fix it
         if (substr_count($content, "\n") > 1) {
             // the final bit of the whitespace must be the next statement's indentation
-            $tokens[$index] = new Token([\T_WHITESPACE, $this->whitespacesConfig->getLineEnding().substr($content, strrpos($content, "\n") + 1)]);
+            $tokens[$index] = new Token([\T_WHITESPACE, $this->whitespacesConfig->getLineEnding().substr($content, (int) strrpos($content, "\n") + 1)]);
         }
     }
 }
